requirements.txt
python-binance==1.0.19
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
ccxt==4.1.13
tensorflow==2.13.0
